+++ packages/common/web/src/index.ts
@@ -1,3 +1,6 @@
+import './types';
+export { ButtonType, ModifierKey, ToolbarType } from './types/toolbarEnums';
+
 // Components
 export { default as AccessibilityListener } from './Components/AccessibilityListener';
 export { default as ViewportRenderer } from './Components/ViewportRenderer';
@@ -48,4 +51,10 @@ export { default as verticalEmbedSchema } from '../statics/schemas/vertical-embe
 export { getImageSrc, WIX_MEDIA_DEFAULT } from './Utils/imageUtils';
 export { isHexColor } from './Utils/colorUtils';
 export { isRtl, getLangDir } from './Utils/rtlUtils';
-export { HEADER_BLOCK, BLOCK_TYPES, HTML_TYPE, SPOILER_TYPE } from './consts';
+export {
+  HEADER_BLOCK,
+  BLOCK_TYPES,
+  HTML_TYPE,
+  SPOILER_TYPE,
+  HEADINGS_DROPDOWN_TYPE,
+} from './consts';